Adding folder structure to tests, minor test refactoring
