+++ spec/features/users/unauthenticated_user_views_pages_spec.rb
