+++ spec/features/users/user_creates_account_spec.rb
