+++ spec/features/users/user_edits_their_profile_spec.rb
