+++ spec/features/users/user_logs_out_spec.rb
