+++ spec/features/users/user_views_profile_spec.rb
